BUG#42370: rpl_trigger.test: Test case timeout after 900 seconds
The test fails with timeout when applying diff on two tables.

In this test case, there can be a situation when the slave is not yet synced
with the master, thence the changes may not be on the slave at diff time.

This patch addresses this issue by synchronizing master and slave before the
diff takes place.
--BZR--
revision-id: luis.soares@sun.com-20090127113216-03o72etze6nzry8a
property-branch-nick: 5.1-rpl
testament3-sha1: 81aa6b169892f7bcb03324f31df8a523bdee3600