BUG#36391: "mysqlbinlog creates invalid charset statements"
The fix for BUG#20103 "Escaping with backslash does not work as expected"
was implemented too greedy though in that it not only changes the behavior
of backslashes within strings but in general, so disabling command shortcuts
like \G or \C (which in turn leads to Bug #36391: "mysqlbinlog creates invalid charset
statements").
      
The fix allows the escaping with backslash to take place only inside a string, 
thus enabling the execution of command shortcuts and presevering the fix for
BUG#20103.
--BZR--
revision-id: alfranio.correia@sun.com-20090127204937-npk0fe90nyuemzmw
property-branch-nick: mysql-5.1-bugteam
testament3-sha1: e38a7bd4f405319a1b100c518ca677b10cd4307c