Bug#42524: Function pthread_setschedprio() is defined but seems broken on i5/OS PASE
The problem is that MySQL use of pthread_setschedprio is not
supported by i5/OS and the default system behavior for unsupported
calls is to emit a SIGILL signal which causes the server to
abort.

The solution is to treat the pthread_setschedprio as inexistent
when compiling binaries for i5/OS. This also does not invalidate
the fix for bug 38477 as the only supported dispatch class is
SCHED_OTHER (which is passed to pthread_setschedparam).
--BZR--
revision-id: davi.arnaut@sun.com-20090206110009-0j96iub8yyxysgk0
property-branch-nick: 42524-5.1
property-file-info: ld7:file_id65:sp1f-configure.in-19700101030959-mgdpoxtnh2ewmvusvfpkreuhwvffkcjw7:message41:Skip pthread_setschedprio check on i5/OS.4:path12:configure.inee
testament3-sha1: a43b0230e4bbc7659d9bdd7b7e894620adc6a6bb