Bug#41110: crash with handler command when used concurrently with alter table Bug#41112: crash in mysql_ha_close_table/get_lock_data with alter table
The problem is that the server wasn't handling robustly failures
to re-open a table during a HANDLER .. READ statement. If the
table needed to be re-opened due to it's storage engine being
altered to one that doesn't support HANDLER, a reference (dangling
pointer) to a closed table could be left in place and accessed in
later attempts to fetch from the table using the handler. Also,
if the server failed to set a error message if the re-open
failed. These problems could lead to server crashes or hangs.

The solution is to remove any references to a closed table and
to set a error if reopening a table during a HANDLER .. READ
statement fails.

There is no test case in this change set as the test depends on
a testing feature only available on 5.1 and later.
--BZR--
revision-id: davi.arnaut@sun.com-20090224091521-frxfa7fg755pih0r
property-branch-nick: 41110-5.0
property-file-info: ld7:file_id67:sp1f-sql_handler.cc-20010406221833-l4tsiortoyipmoyajcoz2tcdppvyeltl7:message126:Remove redundant reopen check.
property-file-info: Set errors even if reopening table.
property-file-info: Reset TABLE_LIST::table reference when the table is closed.4:path18:sql/sql_handler.ccee
testament3-sha1: 1448ae9952e555da65fa0e57bc063ccd45030828