branches/zip
This is to revert the changes made to the plug.in (r4251) as a fix for
issue# 178. Changes to plug.in will not propogate to a plugin
installation unless autotools are rerun which is unacceptable.

A fix for issue# 178 will be committed in a separate commit.

--BZR--
revision-id: svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:4320
testament3-sha1: 943f99e7470d46dc34fcbe1a9601a3941bb2504c