branches/zip:
As suggested by Ken, print a message that says that the Google SMP patch
(GCC atomics) is disabled if it is. Also extend the message when the patch
is partially enabled to make it clear that it is partially enabled.

--BZR--
revision-id: svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:4353
testament3-sha1: 9d6b329806f04932e97b587a2132848920a95d72