Bug #32625: Make test type_bit_innodb more robust
Since there is more than one duplicate value in the table, when adding the
unique index it is not deterministic which value will be reported as causing a
problem.  Replace the reported value with '' so that it doesn't affect the
results.

--BZR--
revision-id: timothy.smith@sun.com-20090311205457-2su41y6hi4oc93dn
property-branch-nick: 50
testament3-sha1: c9ecc588595f547882bdb7b777d05a0ef7993dbb