Bug#36446: Attempt to set @@join_buffer_size to its minimum value produces spurious warning
If a sys-var has a base and a block-size>1, and then a
user-supplied value >= minimum ended up below minimum
thanks to block-size alignment, we threw a warning.
This meant for instance that when getting, then setting
the minimum, we'd see a warning. This was needlessly
confusing. (updated patch)
--BZR--
revision-id: azundris@mysql.com-20090316151145-02w54fu0k8alwxvy
property-branch-nick: 51-36446
testament3-sha1: d9cac0c7ab39298e45e1f7d02bfc7c4c66be9615