This is a backport from 5.1.32 into 5.1.31sp1. The original fix was done by Davi Arnaut on 2009-02-03 All comments are copied from the original fix:
Bug#42524: Function pthread_setschedprio() is defined but seems broken on i5/OS PASE

The problem is that MySQL use of pthread_setschedprio is not
supported i5/OS and the default system behavior for unsupported
calls is to emit a SIGILL signal which causes the server to
abort.

The solution is to treat the pthread_setschedprio as inexistent
when compiling binaries to i5/OS. This also does not invalidate
the fix for bug 38477 as the only supported dispatch class is
SCHED_OTHER (which is passed to pthread_setschedparam).
--BZR--
revision-id: joerg@mysql.com-20090319170947-scyfdstofg9lg1zp
property-branch-nick: qsp-5.1
property-file-info: ld7:file_id65:sp1f-configure.in-19700101030959-mgdpoxtnh2ewmvusvfpkreuhwvffkcjw7:message41:Skip pthread_setschedprio check on i5/OS.4:path12:configure.inee
testament3-sha1: b3e0f17002905e5129b8549fc558a3b60e97ddd2