branches/zip
Enable atomics on solaris (using the libc functions as defined in
atomic.h) if GCC atomic builtins are not present.

There still remains some work to be done (by Vasil?). This patch
makes changes to plug.in to check pthread_t size and presence of
atomic functions when running on solaris. The same has to become
a part of the generated Makefile.in when we bake our source.

Reviewed by: Heikki rb://106

--BZR--
revision-id: svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:4678
testament3-sha1: ecf52ebd389676088a6f485493eac0a11435a627