Bug#39559: dump of stored procedures / functions with C-style \ 	comment can't be read back
A change to the lexer in 5.1 caused slash-asterisk-bang-version
sections to be terminated early if there exists a slash-asterisk-
style comment inside it.  Nesting comments is usually illegal,
but we rely on versioned comment blocks in mysqldump, and the
contents of those sections must be allowed to have comments.

The problem was that when encountering open-comment tokens and
consuming -or- passing through the contents, the "in_comment"
state at the end was clobbered with the not-in-a-comment value,
regardless of whether we were in a comment before this or not.  

So, """/*!VER one /* two */ three */""" would lose its in-comment
state between "two" and "three".  Save the echo and in-comment
state, and restore it at the end of the comment if we consume a 
comment.

--BZR--
revision-id: chad@mysql.com-20090410021818-3w0c8rxeac2idt4g
property-branch-nick: 5.1-bugteam--bug39559
testament3-sha1: ab8093cb8111a7f8e618761b8613a9ada3e744b3