Bug#26887 - main.count_distinct3 fails randomly
Streamlined how we increase the size of our test table.
The new method shows run time decreased by ~60%.
This is not a guarantee that we will not see test timeouts (the random failures noted in the bug),
but it should significantly reduce the chances of this occurring.
--BZR--
revision-id: patrick.crews@sun.com-20090416174744-u9fxu6ophud91a1h
property-branch-nick: mysql-5.0-bugteam
testament3-sha1: 2a2dfde7177ab60813c766e978315b76fb72c107