BUG#36966 - mysqldump.test fails in pushbuild
mysqldump.test is designed to run with concurrent inserts
disabled. It is disabling concurrent inserts at the very
beginning of the test case, and re-enables them at the
bottom of the test. But for some reason (likely incorrect
merge) we enable concurrent inserts in the middle of the test.

The problem is fixed by enabling concurrent inserts only
at the bottom of the test case.
--BZR--
revision-id: svoj@sun.com-20090421090113-hnrvvl029p7gipx4
property-branch-nick: mysql-5.0-bugteam-bug36966
testament3-sha1: a98e924244bb98db5db62e02e49381ab39be2846