Bug #29971 status.test fails
This test uses SHOW STATUS and the like, which may be unstable in the face
of logging to table, since the CSV handler is actively executing operations
and thus incrementing the counters.

Fixed by disabling logging to table for the duration of the test and restoring
it afterwards. This causes various counters to properly start counting from zero
and never advance due to CSV operations.

--BZR--
revision-id: pstoev@mysql.com-20090610085836-2fwbqou6ftta10fb
property-branch-nick: 5.1-bugteam
testament3-sha1: 550523b1a7cfde60931855e2438cd3842ad603a2