Bug #43414 Parenthesis (and other) warnings compiling MySQL with gcc 4.3.2
Compiling MySQL with gcc 4.3.2 and later produces a number of 
warnings, many of which are new with the recent compiler
versions.
            
This bug will be resolved in more than one patch to limit the
size of changesets. This is the second patch, fixing more
of the warnings.
--BZR--
revision-id: staale.smedseng@sun.com-20090610140407-c3aicnb2b32t18an
property-branch-nick: 43414-50
testament3-sha1: 79d21ce95598e23532d596f0678ecca60d577ed6