Bug#42599: error: `pthread_setschedprio' was not declared in this scope
The problem was that a pthread.h header used by gcc did not
declare the pthread_setscheprio, yet the function is implemented
by the function is implemented, causing a autoconf check to pass
and compilation with C++ to fail. The solution is to add a
autoconf check to ensure that the function is properly declared.
--BZR--
revision-id: davi.arnaut@sun.com-20090618232147-i5ytql9whvtnao1z
property-branch-nick: 42599-5.1
property-file-info: ld7:file_id65:sp1f-configure.in-19700101030959-mgdpoxtnh2ewmvusvfpkreuhwvffkcjw7:message48:Check that the pthread_setschedprio is declared.4:path12:configure.inee
testament3-sha1: 09daa486031df755ed07379c84ecdb220ab45710