BUG#44270: Post-push fix
The test case added failed sporadically on PB. This is due to the
fact that the user thread in some cases is waiting for slave IO
to stop and then check the error number. Thence, sometimes the
user thread would race for the error number with IO thread.

This post push fix addresses this by replacing the wait for slave
io to stop with a wait for slave io error (as it seems it was
added in 6.0 also after patch on which this is based was
pushed). This implied backporting wait_for_slave_io_error.inc
from 6.0 also.
--BZR--
revision-id: luis.soares@sun.com-20090626110556-4xkal3zn6l192ytc
property-branch-nick: 5.1-bt-push
testament3-sha1: 43b8e4fe89526f0430ca19c47de1a669216c441a