Bug#45017: Failure to connect if hostname maps to multiple addresses
The problem is that the C API function mysql_real_connect
only attempts to connect to the first IP address returned
for a hostname. This can be a problem if a hostname maps
to multiple IP address and the server is not bound to the
first one that is returned.

The solution is to augment mysql_real_connect so that it
attempts to connect to all IPv4 addresses that a domain
name maps to. The function goes over the list of address
until a successful connection is established.

No test case is provided as its not possible to test this
automatically with the current testing infrastructure.
--BZR--
revision-id: davi.arnaut@sun.com-20090731002843-4ln2597ul9j4v4k0
property-branch-nick: 45017-5.1
property-file-info: ld7:file_id61:sp1f-client.c-20030502160736-oraaciqy6jkfwygs6tqfoaxgjbi65yo77:message175:The client will try to connect to each IPv4 address from
property-file-info: the list of addresses that hostname maps to until a successful
property-file-info: connection is established or there are no more address.4:path19:sql-common/client.cee
property-rebase-of: davi.arnaut@sun.com-20090731002843-y62x3n5qox2hq9zk
testament3-sha1: 100899e177d08e95c5de4e3cbc317b36ccd98fe0