Bug #46019: ERROR 1356 When selecting from within another view that has Group By
Table access rights checking function check_grant() assumed
that no view is opened when it's called.
This is not true with nested views where the inner view
needs materialization. In this case the view is already 
materialized when check_grant() is called for it.
This caused check_grant() to not look for table level
grants on the materialized view table.
Fixed by checking if a view is already materialized and if 
it is check table level grants using the original table name
(not the ones of the materialized temp table).
--BZR--
revision-id: joro@sun.com-20090819121457-7c115y5egff1793a
property-branch-nick: B46019-5.0-bugteam
property-rebase-of: joro@sun.com-20090819121457-2ke3rb1y01pt8mqr
testament3-sha1: 0ea01cc230223092d87b3d349d1fece920b03172